Add Unicode string normalization option (on by default). Pull test vectors from benlaurie/objecthash
Some of the test vectors are not currently working. They are marked with Rspec `xit` to
temporarily disable them but show them in test runs.

All of the existing hexdigest values in docs and specs changed as a result of the
bugfix for issue #2

Added docs to README to show how to turn Unicode normalization off if desired.

Fixed or disabled Rubocop complaints in _spec file