Fixed #45 by allowing hashing of floats.
This was done for almost all implementations in the repository. The only
two exceptions are Java and C. This is because special float values are
not available in JSON and those implementations use JSON objects and not
arbitrary objects.

Also, add ruby to the test suites that the Makefile runs.